Fixes failing tests.
