--- spring-cloud-netflix-core/src/main/java/org/springframework/cloud/netflix/feign/FeignLoggerAutoConfiguration.java
@@ -1,57 +1,0 @@
-/*
- * Copyright 2016 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.cloud.netflix.feign;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import feign.Logger;
-import feign.slf4j.Slf4jLogger;
-
-/**
- * @author Venil Noronha
- */
-@Configuration
-public class FeignLoggerAutoConfiguration {
-
-	@Autowired(required = false)
-	private Logger logger;
-
-	@Bean
-	@ConditionalOnMissingBean(FeignLoggerFactory.class)
-	public FeignLoggerFactory feignLoggerFactory() {
-		return new DefaultFeignLoggerFactory(logger);
-	}
-
-	static class DefaultFeignLoggerFactory implements FeignLoggerFactory {
-
-		private Logger logger;
-
-		public DefaultFeignLoggerFactory(Logger logger) {
-			this.logger = logger;
-		}
-
-		@Override
-		public Logger create(Class<?> type) {
-			return this.logger != null ? this.logger : new Slf4jLogger(type);
-		}
-
-	}
-
-}
